Hot Fix: removed Service folder and moved externalapiservice into Services folder
